+++ frontend/app/src/components/Hoverable.svelte
@@ -134,7 +134,6 @@
 <style lang="scss">
     .noselect {
         @include no_user_select();
-        margin: auto;
 
         &.fill {
             width: 100%;